refactor: /tests/specs → /tests/unit
