+++ tests/index.js
@@ -1,2 +1,2 @@
 require('./integration')
-require('./specs')
+require('./unit')
