+++ tests/unit/is-signed-in-test.js
