+++ tests/unit/sign-out-test.js
