+++ tests/unit/utils-get-properties-test.js
